refactor: domain_modelパッケージを集約ルートでディレクトリ構造に変更
- operational_settings/ディレクトリを作成し、運用設定関連のモデルを配置
- theme_setting/ディレクトリを作成し、テーマ設定関連のモデルを配置
- 各ファイルのimport文を新しいディレクトリ構造に合わせて更新
- 集約ルートパターンに従った構造に変更
