+++ packages/application/lib/operational_settings/check_force_update_use_case.dart
@@ -1,5 +1,5 @@
 import 'package:internal_domain_logic/operational_settings/operational_settings_repository.dart';
-import 'package:internal_domain_model/operating_system.dart';
+import 'package:internal_domain_model/operational_settings/operating_system.dart';
 import 'package:pub_semver/pub_semver.dart';
 
 class CheckForceUpdateUseCase {
