+++ packages/application/lib/operational_settings/check_maintenance_mode_use_case.dart
@@ -1,5 +1,5 @@
 import 'package:internal_domain_logic/operational_settings/operational_settings_repository.dart';
-import 'package:internal_domain_model/operational_settings.dart';
+import 'package:internal_domain_model/operational_settings/operational_settings.dart';
 
 class CheckMaintenanceModeUseCase {
   const CheckMaintenanceModeUseCase({
