+++ packages/domain_logic/lib/operational_settings/operational_settings_repository.dart
@@ -1,4 +1,4 @@
-import 'package:internal_domain_model/operational_settings.dart';
+import 'package:internal_domain_model/operational_settings/operational_settings.dart';
 
 /// 運用設定集約ルートのリポジトリインターフェース
 abstract interface class OperationalSettingsRepository {
