+++ packages/domain_logic/lib/theme_setting/theme_setting_repository.dart
@@ -1,4 +1,4 @@
-import 'package:internal_domain_model/theme_setting.dart';
+import 'package:internal_domain_model/theme_setting/theme_setting.dart';
 
 abstract class ThemeSettingRepository {
   ThemeSetting getThemeSetting();
