+++ packages/domain_model/lib/operational_settings/force_update_settings_state.freezed.dart
