+++ packages/domain_model/lib/operational_settings/maintenance_mode_settings_state.dart
