+++ packages/domain_model/lib/operational_settings/operating_system.dart
