+++ packages/domain_model/lib/operational_settings/operational_settings.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:internal_domain_model/operating_system.dart';
+import 'package:internal_domain_model/operational_settings/operating_system.dart';
 import 'package:pub_semver/pub_semver.dart';
 
 part 'operational_settings.freezed.dart';
