+++ packages/domain_model/lib/operational_settings/operational_settings.freezed.dart
