+++ packages/domain_model/test/operational_settings_test.dart
@@ -1,5 +1,5 @@
-import 'package:internal_domain_model/operating_system.dart';
-import 'package:internal_domain_model/operational_settings.dart';
+import 'package:internal_domain_model/operational_settings/operating_system.dart';
+import 'package:internal_domain_model/operational_settings/operational_settings.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:test/test.dart';
