+++ packages/infrastructure/lib/theme_setting/theme_setting_repository_impl.dart
@@ -1,5 +1,5 @@
 import 'package:internal_domain_logic/theme_setting/theme_setting_repository.dart';
-import 'package:internal_domain_model/theme_setting.dart';
+import 'package:internal_domain_model/theme_setting/theme_setting.dart';
 import 'package:internal_infrastructure/shared_preferences/shared_preference_data_source.dart';
 
 class ThemeSettingRepositoryImpl implements ThemeSettingRepository {
